# deploy/openapi.json
{
  "components": {
    "schemas": {
...
      "EnrollRequest": {
        "additionalProperties": false,
        "properties": {
          "age": {
            "minimum": 0.0,
            "title": "Age",
            "type": "integer"
          },
          "athlete_id": {
            "minLength": 1,
            "title": "Athlete Id",
            "type": "string"
          },
          "bodyweight_kg": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodyweight Kg"
          },
          "experience": {
            "title": "Experience",
            "type": "string"
          },
          "sex": {
            "title": "Sex",
            "type": "string"
          }
        },
        "required": [
          "athlete_id",
          "sex",
          "age",
          "experience"
        ],
        "title": "EnrollRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RevokeRequest": {
        "additionalProperties": false,
        "properties": {
          "token": {
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "RevokeRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Hush v1 \u2014 advisory training intelligence. Hush recommends; the athlete decides.",
    "title": "Hush v1 API",
    "version": "v1",
    "x-hush-schema-version": 11
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/athletes": {
      "post": {
        "description": "Enroll an athlete and mint their bearer token. The athlete is created via the existing\nonboarding seed (ES-008 v2 / Option D); the token's hash is stored, the plaintext returned\nonce for the secure provisioning channel (BB-25).",
        "operationId": "enroll_internal_athletes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enroll",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/athletes/{athlete_id}/erase": {
      "post": {
        "description": "Right-to-erasure (OD-2) \u2014 **logical deletion / anonymization**. Removes the athlete's\nauthentication data + precise personal identifiers and writes an erasure tombstone, while\n**retaining the append-only audit/history in anonymized form** (no destructive audit-chain\ndeletion). The external enrollment/consent record is deleted by Operations as the BB-33\ncomplement. Idempotent; operator-key only.",
        "operationId": "erase_internal_athletes__athlete_id__erase_post",
        "parameters": [
          {
            "in": "path",
            "name": "athlete_id",
            "required": true,
            "schema": {
              "title": "Athlete Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Erase",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/athletes/{athlete_id}/revoke": {
      "post": {
        "operationId": "revoke_internal_athletes__athlete_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "athlete_id",
            "required": true,
            "schema": {
              "title": "Athlete Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/audit/sessions/{session_id}": {
      "get": {
        "description": "Full audit reconstruction (BB-12): composition + recommendation + shadow + observation for\nevery block. A session that cannot be reconstructed is invalid (ES-009 \u00a79).",
        "operationId": "audit_session_internal_audit_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Session",
        "tags": [
          "internal"
        ]
      }
    },
...
  }
}